requirements_dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements_dev.txt pyproject.toml
#
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via
    #   starlette
    #   watchfiles
appium-python-client==1.3.0
    # via pyWinAppDriver (pyproject.toml)
cachetools==5.3.1
    # via tox
chardet==5.2.0
    # via tox
click==8.1.7
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tox
    #   uvicorn
comtypes==1.2.0
    # via pywinauto
distlib==0.3.7
    # via virtualenv
exceptiongroup==1.1.3
    # via
    #   anyio
    #   pytest
fastapi==0.101.1
    # via pyWinAppDriver (pyproject.toml)
filelock==3.12.2
    # via
    #   tox
    #   virtualenv
h11==0.14.0
    # via uvicorn
httptools==0.6.0
    # via uvicorn
idna==3.4
    # via anyio
iniconfig==2.0.0
    # via pytest
lxml==4.9.3
    # via pyWinAppDriver (pyproject.toml)
packaging==23.1
    # via
    #   pyproject-api
    #   pytest
    #   tox
pillow==10.0.0
    # via pyWinAppDriver (pyproject.toml)
platformdirs==3.10.0
    # via
    #   tox
    #   virtualenv
pluggy==1.2.0
    # via
    #   pytest
    #   tox
pydantic==2.2.1
    # via fastapi
pydantic-core==2.6.1
    # via pydantic
pyproject-api==1.5.4
    # via tox
pytest==7.4.0
    # via pyWinAppDriver (pyproject.toml)
python-dotenv==1.0.0
    # via uvicorn
pywin32==306
    # via pywinauto
pywinauto==0.6.8
    # via pyWinAppDriver (pyproject.toml)
pyyaml==6.0.1
    # via uvicorn
selenium==3.141.0
    # via appium-python-client
six==1.16.0
    # via pywinauto
sniffio==1.3.0
    # via anyio
starlette==0.27.0
    # via fastapi
tomli==2.0.1
    # via
    #   pyproject-api
    #   pytest
    #   tox
tox==4.9.0
    # via pyWinAppDriver (pyproject.toml)
typing-extensions==4.7.1
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   uvicorn
urllib3==1.26.16
    # via
    #   pyWinAppDriver (pyproject.toml)
    #   selenium
uvicorn[standard]==0.23.2
    # via pyWinAppDriver (pyproject.toml)
virtualenv==20.24.3
    # via tox
watchfiles==0.19.0
    # via uvicorn
websockets==11.0.3
    # via uvicorn
